fix: refactor db funcs to remove circular dependency
