+++ bot.js
@@ -2,32 +2,12 @@
 const { Client, Collection, Events, GatewayIntentBits } = require('discord.js');
 const fs = require('fs');
 const path = require('path');
-const { MongoClient } = require('mongodb');
+const dbFunc = require('./database/dbFunc')
 
 // Config
 // Token
 require('dotenv').config();
 const token = process.env.DISCORD_TOKEN;
-
-// Database
-const username = process.env.MONGO_DB_USERNAME;
-const password = process.env.MONGO_DB_PASSWORD;
-const db_name = process.env.MONGO_DB_NAME;
-const MONGO_URI = `mongodb+srv://${username}:${password}@${db_name}.dsg1fxk.mongodb.net/?retryWrites=true&w=majority`;
-
-// DB Connection
-const dbClient = new MongoClient(MONGO_URI);
-
-async function connectDatabase() {
-	try {
-		await dbClient.connect();
-		console.log('Connected to database');
-	}
-	catch (error) {
-		dbClient.close();
-		console.error('Error connecting to database: ', error);
-	}
-}
 
 // New Client instance
 const client = new Client({ intents: [GatewayIntentBits.Guilds] });
@@ -91,14 +71,10 @@ client.on(Events.InteractionCreate, async interaction => {
 	}
 });
 
-connectDatabase().then(() => {
+dbFunc.connectDatabase().then(() => {
 	// Log in w/ client token
 	client.login(token);
 }).catch((error) => {
 	console.error('Error: ', error);
 	process.exit(1);
 });
-
-module.exports = {
-	dbClient,
-};
